refactor: 민혜님 리스트와 이름 겹치는 마이아카이브 컴포넌트 파일들 이름 변경
